% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_colors.R
\name{glimpse_colors}
\alias{glimpse_colors}
\title{Glimpse Colors}
\usage{
glimpse_colors(colors_string)
}
\arguments{
\item{colors_string}{}
}
\description{
Glimpse Colors
}
\examples{
glimpse_colors(measure_palette("mixed")(10))
glimpse_colors(measure_palette("primary")(5))
}
