% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_colors.R
\name{measure_colors}
\alias{measure_colors}
\title{Measure Colors}
\usage{
measure_colors(...)
}
\description{
Pulls list of Measure's corporate colors or identifies HEX code of specific color
}
\examples{
measure_colors() Pull list of all colors and HEX codes
measure_colors("Navy") Get specific HEX code
}
