% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_colors.R
\name{measure_palette}
\alias{measure_palette}
\title{Measure Color Palettes}
\usage{
measure_palette(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Options: main, primary, secondary, tertiary, accent, green, gray, dark, mixed}

\item{reverse}{}

\item{...}{}
}
\description{
Generates HEX codes of shades between 2 color palette
}
\examples{
measure_palette("main")(10)
measure_palette("secondary", reverse = TRUE)(10)
}
