% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_colors.R
\name{scale_color_measure}
\alias{scale_color_measure}
\title{Scale Measure Color}
\usage{
scale_color_measure(palette = "main", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Options: main, secondary, tertiary, accent}

\item{discrete}{}

\item{reverse}{}

\item{...}{}
}
\description{
Scale Measure Color
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 4) +
  scale_color_measure(palette = "tertiary", reverse = TRUE)
}
