% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_colors.R
\name{scale_fill_measure}
\alias{scale_fill_measure}
\title{Scale Measure Fill}
\usage{
scale_fill_measure(palette = "main", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Options: main, secondary, tertiary, accent}

\item{discrete}{}

\item{reverse}{}

\item{...}{}
}
\description{
Scale Measure Fill
}
\examples{
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
   geom_point(size = 4, alpha = 0.6) +
   scale_color_measure(discrete = FALSE, palette = "accent", reverse = TRUE)
}
